Replaced AclType enum with integers and constants. Added a binary to initialize an empty DB.

+++ dmlite/trunk/tests/cpp/test-comment.cpp
@@ -16,7 +16,8 @@ public:
 
   void tearDown()
   {
-    this->catalog->unlink(FILE);
+    if (this->catalog)
+      this->catalog->unlink(FILE);
     TestBase::tearDown();
   }
